fix(markdown): preview local links in markdown viewer
