+++ packages/desktop/src/renderer/components/Markdown/index.tsx
@@ -16,16 +16,13 @@ import remarkMath from 'remark-math';
 import 'katex/dist/katex.min.css';
 
 import { openExternalUrl } from '@/renderer/utils/platform';
-import { iconColors } from '@/renderer/styles/colors';
-import { copyText } from '@/renderer/utils/ui/clipboard';
-import { Button, Message, Tooltip } from '@arco-design/web-react';
-import { Copy } from '@icon-park/react';
 import classNames from 'classnames';
 import React, { useCallback, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { convertLatexDelimiters } from '@renderer/utils/chat/latexDelimiters';
 import LocalImageView from '@renderer/components/media/LocalImageView';
 import CodeBlock from './CodeBlock';
+import LocalFileLink from './LocalFileLink';
 import ShadowView from './ShadowView';
 import { resolveLocalFileLinkPath, resolveLocalFileLinkReference } from './markdownUtils';
 import type { LocalFileLinkReference } from './markdownUtils';
@@ -47,88 +44,6 @@ type MarkdownViewProps = {
   onLocalFileLink?: (path: string, reference?: LocalFileLinkReference) => void | Promise<void>;
   /** Enable raw HTML rendering in markdown content. Use with caution — only for trusted sources. */
   allowHtml?: boolean;
-};
-
-const LocalFileLink: React.FC<{
-  reference: LocalFileLinkReference;
-  children?: React.ReactNode;
-  onOpen?: (path: string, reference?: LocalFileLinkReference) => void | Promise<void>;
-}> = ({ reference, children, onOpen }) => {
-  const { t } = useTranslation();
-  const { filePath, line, rawReference } = reference;
-  const fallbackLabel = filePath.split(/[\\/]/).pop() || filePath;
-  const label = children || fallbackLabel;
-  const textLabel =
-    React.Children.toArray(children)
-      .map((child) => (typeof child === 'string' || typeof child === 'number' ? String(child) : ''))
-      .join('') || fallbackLabel;
-  const locationLabel = line == null ? null : `L${line}${reference.column == null ? '' : `:${reference.column}`}`;
-  const canOpen = Boolean(onOpen);
-
-  const handleOpen = useCallback(
-    (event: Event) => {
-      event.preventDefault();
-      event.stopPropagation();
-      if (onOpen) {
-        void onOpen(filePath, reference);
-      }
-    },
-    [filePath, onOpen, reference]
-  );
-
-  const handleCopy = useCallback(
-    (event: Event) => {
-      event.preventDefault();
-      event.stopPropagation();
-      copyText(rawReference).catch(() => {
-        Message.error(t('common.copyFailed'));
-      });
-    },
-    [rawReference, t]
-  );
-
-  return (
-    <span
-      className='inline-flex items-center gap-2px max-w-full align-baseline'
-      data-local-file-path={filePath}
-      data-local-file-line={line}
-      title={rawReference}
-    >
-      {canOpen ? (
-        <Button
-          type='text'
-          size='mini'
-          aria-label={locationLabel ? `${textLabel} ${locationLabel}` : textLabel}
-          className='markdown-local-file-link !px-6px !py-2px !h-auto !leading-normal !align-baseline max-w-full !rd-6px'
-          onClick={handleOpen}
-        >
-          <span className='inline-flex items-center gap-4px max-w-full'>
-            <span className='truncate'>{label}</span>
-            {locationLabel && (
-              <span className='markdown-local-file-line flex-shrink-0 text-11px font-mono'>{locationLabel}</span>
-            )}
-          </span>
-        </Button>
-      ) : (
-        <span className='markdown-local-file-link inline-flex items-center gap-4px max-w-full'>
-          <span className='truncate'>{label}</span>
-          {locationLabel && (
-            <span className='markdown-local-file-line flex-shrink-0 text-11px font-mono'>{locationLabel}</span>
-          )}
-        </span>
-      )}
-      <Tooltip content={t('common.copy', { defaultValue: 'Copy' })}>
-        <Button
-          aria-label={t('common.copy', { defaultValue: 'Copy' })}
-          type='text'
-          size='mini'
-          className='markdown-local-file-copy !p-1px !w-20px !h-20px flex-shrink-0'
-          icon={<Copy theme='outline' size='14' fill={iconColors.secondary} />}
-          onClick={handleCopy}
-        />
-      </Tooltip>
-    </span>
-  );
 };
 
 const MarkdownView: React.FC<MarkdownViewProps> = React.memo(
