+++ packages/desktop/src/renderer/pages/conversation/Preview/hooks/index.ts
@@ -15,3 +15,4 @@ export * from './useTabOverflow';
 export * from './useScrollSync';
 export * from './useScrollSyncHelpers';
 export * from './usePreviewHistory';
+export * from './useLocalFilePreview';
