+++ tests/unit/chat/messageText.dom.test.tsx
@@ -44,7 +44,7 @@ vi.mock('@/common', () => ({
   },
 }));
 
-vi.mock('@/renderer/pages/conversation/Preview', () => ({
+vi.mock('@/renderer/pages/conversation/Preview/context/PreviewContext', () => ({
   usePreviewContext: () => ({
     openPreview: previewMocks.openPreview,
   }),
