+++ tests/unit/previews/MarkdownViewer.dom.test.tsx
@@ -4,15 +4,22 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { describe, it, expect, vi } from 'vitest';
-import { render, screen } from '@testing-library/react';
+import { beforeEach, describe, it, expect, vi } from 'vitest';
+import { fireEvent, render, screen, waitFor } from '@testing-library/react';
 import React from 'react';
+
+const previewMocks = vi.hoisted(() => ({
+  openPreview: vi.fn(),
+}));
+const copyTextMock = vi.hoisted(() => vi.fn().mockResolvedValue(undefined));
 
 vi.mock('@/common', () => ({
   ipcBridge: {
     fs: {
       fetchRemoteImage: { invoke: vi.fn() },
       getImageBase64: { invoke: vi.fn() },
+      getFileMetadata: { invoke: vi.fn() },
+      readFile: { invoke: vi.fn() },
     },
   },
 }));
@@ -40,6 +47,16 @@ vi.mock('@/renderer/utils/platform', () => ({
   openExternalUrl: vi.fn(),
 }));
 
+vi.mock('@/renderer/utils/ui/clipboard', () => ({
+  copyText: copyTextMock,
+}));
+
+vi.mock('@/renderer/pages/conversation/Preview/context/PreviewContext', () => ({
+  usePreviewContext: () => ({
+    openPreview: previewMocks.openPreview,
+  }),
+}));
+
 vi.mock('@/renderer/utils/chat/latexDelimiters', () => ({
   convertLatexDelimiters: (text: string) => text,
 }));
@@ -63,13 +80,52 @@ vi.mock('@/renderer/components/Markdown/MermaidBlock', () => ({
 
 vi.mock('react-i18next', () => ({
   useTranslation: () => ({
-    t: (key: string) => key,
+    t: (key: string, options?: { defaultValue?: string }) => options?.defaultValue ?? key,
   }),
 }));
 
+vi.mock('@arco-design/web-react', () => ({
+  Button: ({
+    children,
+    icon,
+    ...props
+  }: React.ButtonHTMLAttributes<HTMLButtonElement> & { icon?: React.ReactNode }) => (
+    <button type='button' {...props}>
+      {icon}
+      {children}
+    </button>
+  ),
+  Message: {
+    error: vi.fn(),
+  },
+  Tooltip: ({ children }: { children?: React.ReactNode }) => <>{children}</>,
+}));
+
+vi.mock('@icon-park/react', () => ({
+  Copy: () => <span data-testid='copy-icon' />,
+}));
+
 import MarkdownViewer from '@/renderer/pages/conversation/Preview/components/viewers/MarkdownViewer';
+import { ipcBridge } from '@/common';
+
+const fileMetadata = (path: string) => ({
+  name: path.split(/[\\/]/).pop() || path,
+  path,
+  size: 128,
+  type: 'file',
+  lastModified: 1_717_000_000,
+});
 
 describe('MarkdownViewer', () => {
+  beforeEach(() => {
+    previewMocks.openPreview.mockClear();
+    copyTextMock.mockClear();
+    vi.mocked(ipcBridge.fs.getFileMetadata.invoke).mockReset();
+    vi.mocked(ipcBridge.fs.getImageBase64.invoke).mockReset();
+    vi.mocked(ipcBridge.fs.readFile.invoke).mockReset();
+    vi.mocked(ipcBridge.fs.fetchRemoteImage.invoke).mockReset();
+  });
+
   it('renders markdown content in preview mode', () => {
     render(<MarkdownViewer content='# Hello World' />);
     expect(screen.getByText('Hello World')).toBeInTheDocument();
@@ -83,5 +139,52 @@ describe('MarkdownViewer', () => {
   it('hides toolbar when hideToolbar is true', () => {
     render(<MarkdownViewer content='# Test' hideToolbar />);
     expect(screen.queryByText('preview.preview')).not.toBeInTheDocument();
+  });
+
+  it('opens local file links in the preview panel instead of browser windows', async () => {
+    const filePath = '/Users/demo/Desktop/chart.jpg';
+    vi.mocked(ipcBridge.fs.getFileMetadata.invoke).mockResolvedValue(fileMetadata(filePath));
+    vi.mocked(ipcBridge.fs.getImageBase64.invoke).mockResolvedValue('data:image/jpeg;base64,abc123');
+
+    render(<MarkdownViewer content={`[image](${filePath})`} file_path='/Users/demo/Desktop/test.md' />);
+
+    expect(screen.queryByRole('link', { name: 'image' })).not.toBeInTheDocument();
+
+    fireEvent.click(screen.getByRole('button', { name: 'image' }));
+
+    await waitFor(() => {
+      expect(previewMocks.openPreview).toHaveBeenCalledWith(
+        'data:image/jpeg;base64,abc123',
+        'image',
+        expect.objectContaining({
+          file_name: 'chart.jpg',
+          file_path: filePath,
+          language: 'jpg',
+          editable: false,
+        }),
+        { replace: true }
+      );
+    });
+    expect(ipcBridge.fs.getImageBase64.invoke).toHaveBeenCalledWith({ path: filePath, workspace: undefined });
+    expect(ipcBridge.fs.readFile.invoke).not.toHaveBeenCalled();
+  });
+
+  it('keeps remote links as browser anchors', () => {
+    render(<MarkdownViewer content='[docs](https://aionui.com/docs)' />);
+
+    const link = screen.getByRole('link', { name: 'docs' });
+    expect(link).toHaveAttribute('href', 'https://aionui.com/docs');
+  });
+
+  it('continues rendering local image markdown inline', async () => {
+    const filePath = '/Users/demo/Desktop/chart.jpg';
+    vi.mocked(ipcBridge.fs.getImageBase64.invoke).mockResolvedValue('data:image/jpeg;base64,abc123');
+
+    render(<MarkdownViewer content={`![image](${filePath})`} file_path='/Users/demo/Desktop/test.md' />);
+
+    await waitFor(() => {
+      expect(screen.getByRole('img', { name: 'image' })).toHaveAttribute('src', 'data:image/jpeg;base64,abc123');
+    });
+    expect(previewMocks.openPreview).not.toHaveBeenCalled();
   });
 });
